+++ src/main/scala/dev/krysztal/are/foundation/as/As.scala
@@ -1,5 +1,5 @@
 package dev.krysztal.are.foundation.as
 
-trait As[T] {
+trait As[@specialized T] {
     def as(): T
 }